(Rob Lahaye): changes to the xforms graphics dialog.
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@5495 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/insets/insetgraphicsParams.h
@@ -37,7 +37,7 @@ struct InsetGraphicsParams
 	/// How to display the image inside LyX
 	grfx::DisplayType display;
 	/// Scaling for output (LaTeX)
-	unsigned int scale;
+	float scale;
 	/// sizes for output (LaTeX)
 	LyXLength width;
 	///
